refactor: refactor system broker
